bugfix: disable next button for authenticator code generation step when code is still generating
