--- lib/app/features/protect_account/authenticator/views/pages/setup_authenticator/authenticator_code_confirm_page.dart
@@ -1,50 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:ion/app/components/inputs/text_input/components/text_input_icons.dart';
-import 'package:ion/app/components/inputs/text_input/text_input.dart';
-import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
-import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/app/hooks/use_on_init.dart';
-import 'package:ion/generated/assets.gen.dart';
-
-class AuthenticatorCodeConfirmPage extends HookWidget {
-  const AuthenticatorCodeConfirmPage({
-    required this.onFormInitialized,
-    super.key,
-  });
-  final void Function(TextEditingController, GlobalKey<FormState>) onFormInitialized;
-
-  @override
-  Widget build(BuildContext context) {
-    final controller = useTextEditingController();
-    final formKey = useRef(GlobalKey<FormState>());
-
-    useOnInit(
-      () {
-        onFormInitialized(controller, formKey.value);
-      },
-      [],
-    );
-
-    return Form(
-      key: formKey.value,
-      child: Center(
-        child: ScreenSideOffset.large(
-          child: TextInput(
-            controller: controller,
-            labelText: context.i18n.two_fa_auth,
-            prefixIcon: TextInputIcons(
-              hasRightDivider: true,
-              icons: [Assets.svg.iconRecoveryCode.icon()],
-            ),
-            validator: (value) => (value?.isEmpty ?? false) ? '' : null,
-            textInputAction: TextInputAction.done,
-            scrollPadding: EdgeInsets.only(bottom: 200.0.s),
-          ),
-        ),
-      ),
-    );
-  }
-}
