+++ lib/app/features/protect_account/authenticator/views/pages/setup_authenticator/step_pages.dart
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: ice License 1.0
 
-export 'authenticator_code_confirm_page.dart';
-export 'authenticator_insructions_page.dart';
-export 'authenticator_options_page.dart';
-export 'authenticator_success_page.dart';
+export 'authenticator_setup_code_confirm_page.dart';
+export 'authenticator_setup_instructions_page.dart';
+export 'authenticator_setup_options_page.dart';
+export 'authenticator_setup_success_page.dart';
